tried to lint using 'npx eslint --fix src' moved parseCSV to new folder
